requirements.txt
Django==2.2
Pillow
djangorestframework
